Run tests async instead of sync
Since Elixir 1.11 it's been safe to run async tests where you're using
`capture_log` and `capture_io`. The only downside is that - while it is
per process capturing - multiple processes can have their IO or log
output mixed together (only with mixed lines, not any mixes within a
given line), and so assertions on these captures usually need to be
Regexes and not `==` on expected strings.